feat : itempackage 컨트롤러 부분까지

+++ weddingyou_backend/src/main/java/com/mysite/weddingyou_backend/userRegister/UserRegister.java
@@ -1,7 +1,13 @@
 package com.mysite.weddingyou_backend.userRegister;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 
+import com.mysite.weddingyou_backend.item.Item;
+import com.mysite.weddingyou_backend.like.Like;
+
+import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
@@ -9,6 +15,7 @@ import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
@@ -18,10 +25,11 @@ import lombok.Setter;
 @Entity
 @Table(name = "user")
 public class UserRegister {
-    @Id
+    
+	@Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_id")
-    private Long userId;
+    private int userId;
 
     @Column(name = "name", nullable = false)
     private String name;
@@ -43,7 +51,6 @@ public class UserRegister {
 
     @Column(name = "user_join_date", columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP", nullable = false)
     private LocalDateTime userJoinDate;
-    
     
     public enum Gender {
     	male,
